Minor refactoring ahead of upcoming changes.
This just sets up the architecture so we can land subsequent PRs with
cleaner change sets.
